requirements.txt
Django==5.2
django-tailwind==4.0.1
django-crispy-forms==2.4
crispy-tailwind==1.0.3
Pillow==11.2.1
python-dotenv==1.1.0
django-allauth==0.65.7
django-channels==4.2.2
channels-redis==4.2.1
openai==1.75.0
redis==5.2.1
asgiref==3.8.1
channels==4.2.2 